Rename table component + Enable rows limit
