+++ portal/src/app/shared/components/qujata-table/QujataTable.module.scss
