+++ portal/src/app/shared/components/qujata-table/index.ts
@@ -1,0 +1,1 @@
+export * from './QujataTable';
