--- portal/src/app/shared/components/table/index.ts
@@ -1,1 +1,0 @@
-export * from './Table';
